+++ HISTORY.rst
@@ -22,6 +22,7 @@ latest (unreleased)
 **Features and Improvements**
 
 * BSSFL-485: Custom sale order report (add field note)
+* BSSFL-486: Define payment term when sale order is generated from RMA
 * BSSFL-490: On payment, empty the amount when a difference appears
 * BSSFL-493: Install account_asset_management